oracle backend: better support for binding vector<string> parameters
After this change prepare_for_bind just allocates a buffer for binding
without actually copying any data.  The buffer sized based on a size of
the vector and a size of the longest string in it.
Note that these sizes must not be exceeded later.

pre_use now copies the data from the vector<string> into the buffer
allocated in prepare_for_bind.